+++ plugins/woocommerce-admin/packages/components/src/advanced-filters/index.js
@@ -4,11 +4,10 @@
 import { __ } from '@wordpress/i18n';
 import { Component, createRef } from '@wordpress/element';
 import { SelectControl, Button, Dropdown } from '@wordpress/components';
-import { partial, findIndex, difference, isEqual } from 'lodash';
+import { partial, difference, isEqual } from 'lodash';
 import PropTypes from 'prop-types';
 import Gridicon from 'gridicons';
 import interpolateComponents from 'interpolate-components';
-import classnames from 'classnames';
 
 import {
 	getActiveFiltersFromQuery,
@@ -23,11 +22,7 @@ import {
  */
 import Card from '../card';
 import Link from '../link';
-import SelectFilter from './select-filter';
-import SearchFilter from './search-filter';
-import NumberFilter from './number-filter';
-import DateFilter from './date-filter';
-import AttributeFilter from './attribute-filter';
+import AdvancedFilterItem from './item';
 
 const matches = [
 	{ value: 'all', label: __( 'All', 'woocommerce-admin' ) },
@@ -40,9 +35,24 @@ const matches = [
 class AdvancedFilters extends Component {
 	constructor( { query, config } ) {
 		super( ...arguments );
+		this.instanceCounts = {};
+
+		const filtersFromQuery = getActiveFiltersFromQuery(
+			query,
+			config.filters
+		);
+		// @todo: This causes rerenders when instance numbers don't match (from adding/remove before updating query string).
+		const activeFilters = filtersFromQuery.map( ( filter ) => {
+			if ( config.filters[ filter.key ].allowMultiple ) {
+				filter.instance = this.getInstanceNumber( filter.key );
+			}
+
+			return filter;
+		} );
+
 		this.state = {
 			match: query.match || 'all',
-			activeFilters: getActiveFiltersFromQuery( query, config.filters ),
+			activeFilters,
 		};
 
 		this.filterListRef = createRef();
@@ -54,7 +64,6 @@ class AdvancedFilters extends Component {
 		this.removeFilter = this.removeFilter.bind( this );
 		this.clearFilters = this.clearFilters.bind( this );
 		this.getUpdateHref = this.getUpdateHref.bind( this );
-		this.updateFilter = this.updateFilter.bind( this );
 		this.onFilter = this.onFilter.bind( this );
 	}
 
@@ -63,15 +72,34 @@ class AdvancedFilters extends Component {
 		const { query: prevQuery } = prevProps;
 
 		if ( ! isEqual( prevQuery, query ) ) {
-			/* eslint-disable react/no-did-update-set-state */
-			this.setState( {
-				activeFilters: getActiveFiltersFromQuery(
-					query,
-					config.filters
-				),
+			const filtersFromQuery = getActiveFiltersFromQuery(
+				query,
+				config.filters
+			);
+
+			// Update all multiple instance counts.
+			this.instanceCounts = {};
+			// @todo: This causes rerenders when instance numbers don't match (from adding/remove before updating query string).
+			const activeFilters = filtersFromQuery.map( ( filter ) => {
+				if ( config.filters[ filter.key ].allowMultiple ) {
+					filter.instance = this.getInstanceNumber( filter.key );
+				}
+
+				return filter;
 			} );
+
+			/* eslint-disable react/no-did-update-set-state */
+			this.setState( { activeFilters } );
 			/* eslint-enable react/no-did-update-set-state */
 		}
+	}
+
+	getInstanceNumber( key ) {
+		if ( ! this.instanceCounts.hasOwnProperty( key ) ) {
+			this.instanceCounts[ key ] = 1;
+		}
+
+		return this.instanceCounts[ key ]++;
 	}
 
 	onMatchChange( match ) {
@@ -82,41 +110,19 @@ class AdvancedFilters extends Component {
 		onAdvancedFilterAction( 'match', { match } );
 	}
 
-	onFilterChange( key, property, value ) {
-		const activeFilters = this.state.activeFilters.map(
-			( activeFilter ) => {
-				if ( key === activeFilter.key ) {
-					return Object.assign( {}, activeFilter, {
-						[ property ]: value,
-					} );
-				}
-				return activeFilter;
-			}
-		);
+	onFilterChange( index, property, value ) {
+		const newActiveFilters = [ ...this.state.activeFilters ];
+		newActiveFilters[ index ] = {
+			...newActiveFilters[ index ],
+			[ property ]: value,
+		};
 
-		this.setState( { activeFilters } );
+		this.setState( { activeFilters: newActiveFilters } );
 	}
 
-	updateFilter( filter ) {
-		const activeFilters = this.state.activeFilters.map(
-			( activeFilter ) => {
-				if ( filter.key === activeFilter.key ) {
-					return filter;
-				}
-				return activeFilter;
-			}
-		);
-
-		this.setState( { activeFilters } );
-	}
-
-	removeFilter( key ) {
+	removeFilter( index ) {
 		const { onAdvancedFilterAction } = this.props;
 		const activeFilters = [ ...this.state.activeFilters ];
-		const index = findIndex(
-			activeFilters,
-			( filter ) => filter.key === key
-		);
 		onAdvancedFilterAction( 'remove', activeFilters[ index ] );
 		activeFilters.splice( index, 1 );
 		this.setState( { activeFilters } );
@@ -151,7 +157,17 @@ class AdvancedFilters extends Component {
 	getAvailableFilterKeys() {
 		const { config } = this.props;
 		const activeFilterKeys = this.state.activeFilters.map( ( f ) => f.key );
-		return difference( Object.keys( config.filters ), activeFilterKeys );
+		const multipleValueFilterKeys = Object.keys( config.filters ).filter(
+			( f ) => config.filters[ f ].allowMultiple || false
+		);
+		const inactiveFilterKeys = difference(
+			Object.keys( config.filters ),
+			activeFilterKeys,
+			multipleValueFilterKeys
+		);
+
+		// Ensure filters that allow multiples are alway present.
+		return [ ...inactiveFilterKeys, ...multipleValueFilterKeys ];
 	}
 
 	addFilter( key, onClose ) {
@@ -172,6 +188,9 @@ class AdvancedFilters extends Component {
 		}
 		if ( filterConfig.input && filterConfig.input.component === 'Search' ) {
 			newFilter.value = '';
+		}
+		if ( filterConfig.allowMultiple ) {
+			newFilter.instance = this.getInstanceNumber( key );
 		}
 		this.setState( ( state ) => {
 			return {
@@ -256,112 +275,24 @@ class AdvancedFilters extends Component {
 				>
 					{ activeFilters
 						.sort( this.orderFilters )
-						.map( ( filter ) => {
-							const { key } = filter;
-							const { input, labels } = config.filters[ key ];
+						.map( ( filter, idx ) => {
+							const { instance, key } = filter;
 							return (
-								<li
-									className="woocommerce-filters-advanced__list-item"
-									key={ key }
-								>
-									{ input.component === 'SelectControl' && (
-										<SelectFilter
-											className="woocommerce-filters-advanced__fieldset-item"
-											filter={ filter }
-											config={ config.filters[ key ] }
-											onFilterChange={
-												this.onFilterChange
-											}
-											isEnglish={ isEnglish }
-										/>
+								<AdvancedFilterItem
+									key={ key + ( instance || '' ) }
+									config={ config }
+									currency={ currency }
+									filter={ filter }
+									isEnglish={ isEnglish }
+									onFilterChange={ partial(
+										this.onFilterChange,
+										idx
 									) }
-									{ input.component === 'Search' && (
-										<SearchFilter
-											className="woocommerce-filters-advanced__fieldset-item"
-											filter={ filter }
-											config={ config.filters[ key ] }
-											onFilterChange={
-												this.onFilterChange
-											}
-											isEnglish={ isEnglish }
-											query={ query }
-										/>
-									) }
-									{ input.component === 'Number' && (
-										<NumberFilter
-											className="woocommerce-filters-advanced__fieldset-item"
-											filter={ filter }
-											config={ config.filters[ key ] }
-											onFilterChange={
-												this.onFilterChange
-											}
-											isEnglish={ isEnglish }
-											query={ query }
-											currency={ currency }
-										/>
-									) }
-									{ input.component === 'Currency' && (
-										<NumberFilter
-											className="woocommerce-filters-advanced__fieldset-item"
-											filter={ filter }
-											config={ {
-												...config.filters[ key ],
-												...{
-													input: {
-														type: 'currency',
-														component: 'Currency',
-													},
-												},
-											} }
-											onFilterChange={
-												this.onFilterChange
-											}
-											isEnglish={ isEnglish }
-											query={ query }
-											currency={ currency }
-										/>
-									) }
-									{ input.component === 'Date' && (
-										<DateFilter
-											className="woocommerce-filters-advanced__fieldset-item"
-											filter={ filter }
-											config={ config.filters[ key ] }
-											onFilterChange={
-												this.onFilterChange
-											}
-											isEnglish={ isEnglish }
-											query={ query }
-											updateFilter={ this.updateFilter }
-										/>
-									) }
-									{ input.component ===
-										'ProductAttribute' && (
-										<AttributeFilter
-											className="woocommerce-filters-advanced__fieldset-item"
-											filter={ filter }
-											config={ config.filters[ key ] }
-											onFilterChange={
-												this.onFilterChange
-											}
-											isEnglish={ isEnglish }
-											query={ query }
-											updateFilter={ this.updateFilter }
-										/>
-									) }
-									<Button
-										className={ classnames(
-											'woocommerce-filters-advanced__line-item',
-											'woocommerce-filters-advanced__remove'
-										) }
-										label={ labels.remove }
-										onClick={ partial(
-											this.removeFilter,
-											key
-										) }
-									>
-										<Gridicon icon="cross-small" />
-									</Button>
-								</li>
+									query={ query }
+									removeFilter={ () =>
+										this.removeFilter( idx )
+									}
+								/>
 							);
 						} ) }
 				</ul>