+++ plugins/woocommerce-admin/packages/components/src/advanced-filters/number-filter.js
@@ -137,7 +137,7 @@ class NumberFilter extends Component {
 		if ( Boolean( rangeEnd ) ) {
 			// If there's a value for rangeEnd, we've just changed from "between"
 			// to "less than" or "more than" and need to transition the value
-			onFilterChange( filter.key, 'value', rangeStart || rangeEnd );
+			onFilterChange( 'value', rangeStart || rangeEnd );
 		}
 
 		let labelFormat = '';
@@ -166,7 +166,7 @@ class NumberFilter extends Component {
 			label: sprintf( labelFormat, {
 				field: get( config, [ 'labels', 'add' ] ),
 			} ),
-			onChange: partial( onFilterChange, filter.key, 'value' ),
+			onChange: partial( onFilterChange, 'value' ),
 			currencySymbol,
 			symbolPosition,
 		} );
@@ -181,11 +181,11 @@ class NumberFilter extends Component {
 			: [ filter.value ];
 
 		const rangeStartOnChange = ( newRangeStart ) => {
-			onFilterChange( filter.key, 'value', [ newRangeStart, rangeEnd ] );
+			onFilterChange( 'value', [ newRangeStart, rangeEnd ] );
 		};
 
 		const rangeEndOnChange = ( newRangeEnd ) => {
-			onFilterChange( filter.key, 'value', [ rangeStart, newRangeEnd ] );
+			onFilterChange( 'value', [ rangeStart, newRangeEnd ] );
 		};
 
 		return interpolateComponents( {
@@ -230,7 +230,7 @@ class NumberFilter extends Component {
 			onFilterChange,
 			isEnglish,
 		} = this.props;
-		const { key, rule } = filter;
+		const { rule } = filter;
 		const { labels, rules } = config;
 
 		const children = interpolateComponents( {
@@ -245,7 +245,7 @@ class NumberFilter extends Component {
 						) }
 						options={ rules }
 						value={ rule }
-						onChange={ partial( onFilterChange, key, 'rule' ) }
+						onChange={ partial( onFilterChange, 'rule' ) }
 						aria-label={ labels.rule }
 					/>
 				),